Separación de los templates home, sign in, sign up en diferentes archivos ya que estos se encontraban en un solo archivo js
